LineageEditableGridLoaderFromSelection: process columns internally
