+++ packages/components/src/index.ts
@@ -306,7 +306,6 @@ import {
     deleteSampleSet,
     fetchSamples,
     getFieldLookupFromSelection,
-    getLineageEditorUpdateColumns,
     getSampleSet,
     getSampleTypeDetails,
     getSelectionLineageData,
@@ -1007,7 +1006,6 @@ export {
     getFieldLookupFromSelection,
     getSelectedItemSamples,
     getSelectionLineageData,
-    getLineageEditorUpdateColumns,
     EntityInsertPanel,
     EntityCrossProjectSelectionConfirmModal,
     SampleTypeDataType,
