+++ packages/components/src/internal/components/samples/actions.spec.ts
@@ -1,14 +1,6 @@
 import { List, fromJS } from 'immutable';
 
-import { QueryInfo } from '../../../public/QueryInfo';
-import { makeTestQueryModel } from '../../../public/QueryModel/testUtils';
-import { SchemaQuery } from '../../../public/SchemaQuery';
-import { DataClassDataType, SampleTypeDataType } from '../entities/constants';
-import { EntityChoice, IEntityTypeOption } from '../entities/models';
-
-import { QueryColumn } from '../../../public/QueryColumn';
-
-import { getLineageEditorUpdateColumns, getUpdatedLineageRows, getRowIdsFromSelection } from './actions';
+import { getUpdatedLineageRows, getRowIdsFromSelection } from './actions';
 
 let DATA = fromJS({
     '1': {
@@ -126,87 +118,6 @@ describe('getUpdatedLineageRows', () => {
             [1, 2]
         );
         expect(updatedRows.length).toBe(0);
-    });
-});
-
-describe('getLineageEditorUpdateColumns', () => {
-    const MODEL = makeTestQueryModel(
-        SchemaQuery.create('schema', 'query'),
-        QueryInfo.fromJSON({
-            columns: [
-                { fieldKey: 'rowId' },
-                {
-                    fieldKey: 'name',
-                    fieldKeyArray: ['name'],
-                    shownInUpdateView: true,
-                    userEditable: true,
-                },
-                { fieldKey: 'other' },
-            ],
-        })
-    );
-
-    test('no parent types', () => {
-        const cols = getLineageEditorUpdateColumns(MODEL, {});
-        expect(cols.queryInfoColumns.size).toBe(2);
-        expect(cols.queryInfoColumns.get('rowid')).toBeDefined();
-        expect(cols.queryInfoColumns.get('name')).toBeDefined();
-        expect(cols.queryInfoColumns.get('other')).toBeUndefined();
-        expect(cols.updateColumns.size).toBe(1);
-        expect(cols.updateColumns.get(0).get('fieldKey')).toBe('name');
-    });
-
-    test('with parent types', () => {
-        const cols = getLineageEditorUpdateColumns(MODEL, {
-            s1: List.of({
-                type: {
-                    lsid: 'a',
-                    rowId: 1,
-                    schema: 'exp',
-                    query: 'test1',
-                    entityDataType: SampleTypeDataType,
-                } as IEntityTypeOption,
-                ids: [],
-                value: '',
-            } as EntityChoice),
-            s2: List.of({
-                type: {
-                    lsid: 'b',
-                    rowId: 1,
-                    schema: 'exp.data',
-                    query: 'test2',
-                    entityDataType: DataClassDataType,
-                } as IEntityTypeOption,
-                ids: [],
-                value: '',
-            } as EntityChoice),
-            s3: List.of(),
-        });
-        expect(cols.queryInfoColumns.size).toBe(4);
-        expect(cols.queryInfoColumns.get('rowid')).toBeDefined();
-        expect(cols.queryInfoColumns.get('name')).toBeDefined();
-        expect(cols.queryInfoColumns.get('other')).toBeUndefined();
-        expect(cols.queryInfoColumns.get('MaterialInputs/Test1')).toBeDefined();
-        expect(cols.queryInfoColumns.get('DataInputs/Test2')).toBeDefined();
-        expect(cols.updateColumns.size).toBe(3);
-        expect(cols.updateColumns.get(0).get('fieldKey')).toBe('name');
-        expect(cols.updateColumns.get(1).get('fieldKey')).toBe('DataInputs/Test2');
-        expect(cols.updateColumns.get(2).get('fieldKey')).toBe('MaterialInputs/Test1');
-    });
-
-    // Regression coverage to ensure the "name" column is included in the lineage
-    // update columns ONLY when the "name" column is a part of the update columns for the underlying QueryInfo.
-    test('without updatable "name" column', () => {
-        const queryInfo = MODEL.queryInfo.setIn(
-            ['columns', 'name'],
-            new QueryColumn({ fieldKey: 'name' })
-        ) as QueryInfo;
-        const model = MODEL.mutate({ queryInfo });
-        const cols = getLineageEditorUpdateColumns(model, {});
-
-        expect(cols.queryInfoColumns.size).toBe(2);
-        expect(cols.queryInfoColumns.get('name')).toBeDefined();
-        expect(cols.updateColumns.size).toBe(0);
     });
 });
